requirements.txt
streamlit==1.24.0
cryptography>=40.0.0
deep-translator>=1.9.1
langdetect>=1.0.9
textblob>=0.17.1
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
